feat: move finances to the client
